refactor(lint): fix react-refresh and hook-deps warnings, run formatter
- Move pending-invitation token helpers out of AcceptInvitationModal into
  src/utils/pendingInvitationToken.ts so the component file only exports
  components (react-refresh/only-export-components)
- Memoize the invitations fallback in ShareNoteDialog so its callback
  dependencies are stable (react-hooks/exhaustive-deps)
- Prettier reflow in clipboard.js
