+++ src/utils/pendingInvitationToken.ts
@@ -1,0 +1,16 @@
+const STORAGE_KEY = "pendingInvitationToken";
+
+export function storePendingInvitationToken(token: string): void {
+  if (typeof window === "undefined") return;
+  localStorage.setItem(STORAGE_KEY, token);
+}
+
+export function consumePendingInvitationToken(): string | null {
+  if (typeof window === "undefined") return null;
+  return localStorage.getItem(STORAGE_KEY);
+}
+
+export function clearPendingInvitationToken(): void {
+  if (typeof window === "undefined") return;
+  localStorage.removeItem(STORAGE_KEY);
+}
